[fix](test) disambiguate scanner counter values

+++ be/test/scan/scanner_context_test.cpp
@@ -691,9 +691,9 @@ TEST_F(ScannerContextTest, get_free_block) {
             state.get(), olap_scan_local_state.get(), output_tuple_desc, output_row_descriptor,
             scanners, limit, scan_dependency, parallel_tasks);
     scanner_context->_newly_create_free_blocks_num = newly_create_free_blocks_num.get();
-    scanner_context->_newly_create_free_blocks_num->set(0L);
+    scanner_context->_newly_create_free_blocks_num->set(int64_t(0));
     scanner_context->_scanner_memory_used_counter = scanner_memory_used_counter.get();
-    scanner_context->_scanner_memory_used_counter->set(0L);
+    scanner_context->_scanner_memory_used_counter->set(int64_t(0));
     BlockUPtr block = scanner_context->get_free_block(/*force=*/true);
     ASSERT_NE(block, nullptr);
     ASSERT_TRUE(scanner_context->_newly_create_free_blocks_num->value() == 1);
